daneel/ai/requirements.txt
serial
bitstring
smbus
pyyaml
